Fix volunteer attendance page and schedule display
- Rename "Manage Volunteers" to "Volunteer Attendance"
- Change hours input to accept only integers (remove step="0.5")
- Fix database connection error for event_participation table (use sashvini)
- Fix event_role queries to use izzati connection
- Fix schedule page showing decimal days (cast to int)

+++ resources/views/volunteer-management/schedule.blade.php
@@ -268,7 +268,7 @@
                                                 <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z"></path>
                                             </svg>
                                             @php
-                                                $daysUntil = now()->diffInDays($event->Start_Date, false);
+                                                $daysUntil = (int) now()->diffInDays($event->Start_Date, false);
                                             @endphp
                                             <span class="
                                                 @if($daysUntil == 0) text-red-600
